Disabled Python's output buffering.
Common problem of dockerized python apps is that we don't see any
logs. The problem is Python buffers stdout and stderr, which makes
it look like there are no logs. In reality, they are delayed.

Since we want to capture everything immediately, it makes sense
for us to turn off buffering.