The producer script now uses KafkaProducer.
SimpleProducer is not only a deprecated API, it is problematic
when it comes to handling disconnects and retries.

The KafkaProducer, while harder to setup, offers more features,
including handling retries.

The config I put here did not come from me. This was copied from
another project which I know is working. For now, we'll stick
to this. We can make this configurable later on.